create dedicated file for the logo + handle success error login
