+++ ui/app/src/components/logo/LightThemePersesLogo.tsx
@@ -1,0 +1,48 @@
+// Copyright 2023 The Perses Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+function LightThemePersesLogo() {
+  return (
+    <svg width="394" height="116" viewBox="0 0 394 116" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clipPath="url(#clip0_2_9)">
+        <path
+          d="M141.76 80V35.328H158.016C162.155 35.328 165.312 36.459 167.488 38.72C169.664 40.981 170.752 44.139 170.752 48.192C170.752 52.245 169.664 55.403 167.488 57.664C165.312 59.925 162.155 61.056 158.016 61.056H147.136V80H141.76ZM147.136 56.384H157.888C160.149 56.384 161.899 55.829 163.136 54.72C164.416 53.568 165.056 51.947 165.056 49.856V46.528C165.056 44.437 164.416 42.837 163.136 41.728C161.899 40.576 160.149 40 157.888 40H147.136V56.384ZM180.135 80V35.328H207.655V40H185.511V55.04H206.887V59.712H185.511V75.328H207.655V80H180.135ZM223.886 80H218.51V35.328H234.766C238.905 35.328 242.062 36.459 244.238 38.72C246.414 40.981 247.502 44.139 247.502 48.192C247.502 51.819 246.563 54.72 244.686 56.896C242.809 59.072 240.078 60.331 236.494 60.672L247.886 80H241.87L230.99 60.928H223.886V80ZM234.638 56.384C236.899 56.384 238.649 55.829 239.886 54.72C241.166 53.568 241.806 51.947 241.806 49.856V46.528C241.806 44.437 241.166 42.837 239.886 41.728C238.649 40.576 236.899 40 234.638 40H223.886V56.384H234.638ZM270.197 80.768C266.186 80.768 262.901 80.043 260.341 78.592C257.781 77.099 255.669 75.2 254.005 72.896L257.909 69.696C259.658 71.872 261.493 73.493 263.413 74.56C265.376 75.627 267.701 76.16 270.389 76.16C273.632 76.16 276.106 75.413 277.813 73.92C279.52 72.427 280.373 70.336 280.373 67.648C280.373 65.472 279.754 63.787 278.517 62.592C277.28 61.355 275.125 60.48 272.053 59.968L267.061 59.136C264.885 58.752 263.05 58.176 261.557 57.408C260.106 56.64 258.933 55.723 258.037 54.656C257.141 53.547 256.501 52.352 256.117 51.072C255.733 49.749 255.541 48.384 255.541 46.976C255.541 42.88 256.885 39.787 259.573 37.696C262.261 35.605 265.888 34.56 270.453 34.56C274.037 34.56 277.045 35.157 279.477 36.352C281.952 37.547 283.936 39.232 285.429 41.408L281.653 44.672C280.416 43.051 278.922 41.728 277.173 40.704C275.424 39.68 273.162 39.168 270.389 39.168C267.36 39.168 265.034 39.808 263.413 41.088C261.792 42.368 260.981 44.288 260.981 46.848C260.981 48.853 261.578 50.496 262.773 51.776C264.01 53.013 266.208 53.909 269.365 54.464L274.165 55.296C276.341 55.68 278.176 56.256 279.669 57.024C281.162 57.792 282.357 58.709 283.253 59.776C284.192 60.843 284.853 62.037 285.237 63.36C285.621 64.683 285.813 66.069 285.813 67.52C285.813 71.659 284.448 74.901 281.717 77.248C279.029 79.595 275.189 80.768 270.197 80.768ZM295.26 80V35.328H322.78V40H300.636V55.04H322.012V59.712H300.636V75.328H322.78V80H295.26ZM346.947 80.768C342.936 80.768 339.651 80.043 337.091 78.592C334.531 77.099 332.419 75.2 330.755 72.896L334.659 69.696C336.408 71.872 338.243 73.493 340.163 74.56C342.126 75.627 344.451 76.16 347.139 76.16C350.382 76.16 352.856 75.413 354.563 73.92C356.27 72.427 357.123 70.336 357.123 67.648C357.123 65.472 356.504 63.787 355.267 62.592C354.03 61.355 351.875 60.48 348.803 59.968L343.811 59.136C341.635 58.752 339.8 58.176 338.307 57.408C336.856 56.64 335.683 55.723 334.787 54.656C333.891 53.547 333.251 52.352 332.867 51.072C332.483 49.749 332.291 48.384 332.291 46.976C332.291 42.88 333.635 39.787 336.323 37.696C339.011 35.605 342.638 34.56 347.203 34.56C350.787 34.56 353.795 35.157 356.227 36.352C358.702 37.547 360.686 39.232 362.179 41.408L358.403 44.672C357.166 43.051 355.672 41.728 353.923 40.704C352.174 39.68 349.912 39.168 347.139 39.168C344.11 39.168 341.784 39.808 340.163 41.088C338.542 42.368 337.731 44.288 337.731 46.848C337.731 48.853 338.328 50.496 339.523 51.776C340.76 53.013 342.958 53.909 346.115 54.464L350.915 55.296C353.091 55.68 354.926 56.256 356.419 57.024C357.912 57.792 359.107 58.709 360.003 59.776C360.942 60.843 361.603 62.037 361.987 63.36C362.371 64.683 362.563 66.069 362.563 67.52C362.563 71.659 361.198 74.901 358.467 77.248C355.779 79.595 351.939 80.768 346.947 80.768Z"
+          fill="#15191F"
+        />
+        <path
+          d="M86.625 18H44.375C40.8542 18 38 20.8542 38 24.375C38 27.8958 40.8542 30.75 44.375 30.75H86.625C90.1458 30.75 93 27.8958 93 24.375C93 20.8542 90.1458 18 86.625 18Z"
+          fill="#DE005D"
+        />
+        <path
+          d="M106.625 38.75H64.375C60.8542 38.75 58 41.6042 58 45.125C58 48.6458 60.8542 51.5 64.375 51.5H106.625C110.146 51.5 113 48.6458 113 45.125C113 41.6042 110.146 38.75 106.625 38.75Z"
+          fill="#DE005D"
+        />
+        <path
+          d="M86.625 59.5H44.375C40.8542 59.5 38 62.3542 38 65.875C38 69.3958 40.8542 72.25 44.375 72.25H86.625C90.1458 72.25 93 69.3958 93 65.875C93 62.3542 90.1458 59.5 86.625 59.5Z"
+          fill="#DE005D"
+        />
+        <path
+          d="M51.625 80.25H44.375C40.8542 80.25 38 83.1042 38 86.625C38 90.1458 40.8542 93 44.375 93H51.625C55.1458 93 58 90.1458 58 86.625C58 83.1042 55.1458 80.25 51.625 80.25Z"
+          fill="#DE005D"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_2_9">
+          <rect width="394" height="116" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export default LightThemePersesLogo;
