+++ ui/app/src/views/auth/AuthView.tsx
@@ -1,99 +1,46 @@
+// Copyright 2023 The Perses Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import { Button, Container, Divider, LinearProgress, TextField, useMediaQuery, useTheme } from '@mui/material';
+import { useState } from 'react';
+import { useSnackbar } from '@perses-dev/components';
 import { useDarkMode } from '../../context/DarkMode';
 import { useAuthMutation } from '../../model/auth-client';
-import { useState } from 'react';
-
-function LightThemePersesLogo() {
-  return (
-    <svg width="394" height="116" viewBox="0 0 394 116" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clipPath="url(#clip0_2_9)">
-        <path
-          d="M141.76 80V35.328H158.016C162.155 35.328 165.312 36.459 167.488 38.72C169.664 40.981 170.752 44.139 170.752 48.192C170.752 52.245 169.664 55.403 167.488 57.664C165.312 59.925 162.155 61.056 158.016 61.056H147.136V80H141.76ZM147.136 56.384H157.888C160.149 56.384 161.899 55.829 163.136 54.72C164.416 53.568 165.056 51.947 165.056 49.856V46.528C165.056 44.437 164.416 42.837 163.136 41.728C161.899 40.576 160.149 40 157.888 40H147.136V56.384ZM180.135 80V35.328H207.655V40H185.511V55.04H206.887V59.712H185.511V75.328H207.655V80H180.135ZM223.886 80H218.51V35.328H234.766C238.905 35.328 242.062 36.459 244.238 38.72C246.414 40.981 247.502 44.139 247.502 48.192C247.502 51.819 246.563 54.72 244.686 56.896C242.809 59.072 240.078 60.331 236.494 60.672L247.886 80H241.87L230.99 60.928H223.886V80ZM234.638 56.384C236.899 56.384 238.649 55.829 239.886 54.72C241.166 53.568 241.806 51.947 241.806 49.856V46.528C241.806 44.437 241.166 42.837 239.886 41.728C238.649 40.576 236.899 40 234.638 40H223.886V56.384H234.638ZM270.197 80.768C266.186 80.768 262.901 80.043 260.341 78.592C257.781 77.099 255.669 75.2 254.005 72.896L257.909 69.696C259.658 71.872 261.493 73.493 263.413 74.56C265.376 75.627 267.701 76.16 270.389 76.16C273.632 76.16 276.106 75.413 277.813 73.92C279.52 72.427 280.373 70.336 280.373 67.648C280.373 65.472 279.754 63.787 278.517 62.592C277.28 61.355 275.125 60.48 272.053 59.968L267.061 59.136C264.885 58.752 263.05 58.176 261.557 57.408C260.106 56.64 258.933 55.723 258.037 54.656C257.141 53.547 256.501 52.352 256.117 51.072C255.733 49.749 255.541 48.384 255.541 46.976C255.541 42.88 256.885 39.787 259.573 37.696C262.261 35.605 265.888 34.56 270.453 34.56C274.037 34.56 277.045 35.157 279.477 36.352C281.952 37.547 283.936 39.232 285.429 41.408L281.653 44.672C280.416 43.051 278.922 41.728 277.173 40.704C275.424 39.68 273.162 39.168 270.389 39.168C267.36 39.168 265.034 39.808 263.413 41.088C261.792 42.368 260.981 44.288 260.981 46.848C260.981 48.853 261.578 50.496 262.773 51.776C264.01 53.013 266.208 53.909 269.365 54.464L274.165 55.296C276.341 55.68 278.176 56.256 279.669 57.024C281.162 57.792 282.357 58.709 283.253 59.776C284.192 60.843 284.853 62.037 285.237 63.36C285.621 64.683 285.813 66.069 285.813 67.52C285.813 71.659 284.448 74.901 281.717 77.248C279.029 79.595 275.189 80.768 270.197 80.768ZM295.26 80V35.328H322.78V40H300.636V55.04H322.012V59.712H300.636V75.328H322.78V80H295.26ZM346.947 80.768C342.936 80.768 339.651 80.043 337.091 78.592C334.531 77.099 332.419 75.2 330.755 72.896L334.659 69.696C336.408 71.872 338.243 73.493 340.163 74.56C342.126 75.627 344.451 76.16 347.139 76.16C350.382 76.16 352.856 75.413 354.563 73.92C356.27 72.427 357.123 70.336 357.123 67.648C357.123 65.472 356.504 63.787 355.267 62.592C354.03 61.355 351.875 60.48 348.803 59.968L343.811 59.136C341.635 58.752 339.8 58.176 338.307 57.408C336.856 56.64 335.683 55.723 334.787 54.656C333.891 53.547 333.251 52.352 332.867 51.072C332.483 49.749 332.291 48.384 332.291 46.976C332.291 42.88 333.635 39.787 336.323 37.696C339.011 35.605 342.638 34.56 347.203 34.56C350.787 34.56 353.795 35.157 356.227 36.352C358.702 37.547 360.686 39.232 362.179 41.408L358.403 44.672C357.166 43.051 355.672 41.728 353.923 40.704C352.174 39.68 349.912 39.168 347.139 39.168C344.11 39.168 341.784 39.808 340.163 41.088C338.542 42.368 337.731 44.288 337.731 46.848C337.731 48.853 338.328 50.496 339.523 51.776C340.76 53.013 342.958 53.909 346.115 54.464L350.915 55.296C353.091 55.68 354.926 56.256 356.419 57.024C357.912 57.792 359.107 58.709 360.003 59.776C360.942 60.843 361.603 62.037 361.987 63.36C362.371 64.683 362.563 66.069 362.563 67.52C362.563 71.659 361.198 74.901 358.467 77.248C355.779 79.595 351.939 80.768 346.947 80.768Z"
-          fill="#15191F"
-        />
-        <path
-          d="M86.625 18H44.375C40.8542 18 38 20.8542 38 24.375C38 27.8958 40.8542 30.75 44.375 30.75H86.625C90.1458 30.75 93 27.8958 93 24.375C93 20.8542 90.1458 18 86.625 18Z"
-          fill="#DE005D"
-        />
-        <path
-          d="M106.625 38.75H64.375C60.8542 38.75 58 41.6042 58 45.125C58 48.6458 60.8542 51.5 64.375 51.5H106.625C110.146 51.5 113 48.6458 113 45.125C113 41.6042 110.146 38.75 106.625 38.75Z"
-          fill="#DE005D"
-        />
-        <path
-          d="M86.625 59.5H44.375C40.8542 59.5 38 62.3542 38 65.875C38 69.3958 40.8542 72.25 44.375 72.25H86.625C90.1458 72.25 93 69.3958 93 65.875C93 62.3542 90.1458 59.5 86.625 59.5Z"
-          fill="#DE005D"
-        />
-        <path
-          d="M51.625 80.25H44.375C40.8542 80.25 38 83.1042 38 86.625C38 90.1458 40.8542 93 44.375 93H51.625C55.1458 93 58 90.1458 58 86.625C58 83.1042 55.1458 80.25 51.625 80.25Z"
-          fill="#DE005D"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_2_9">
-          <rect width="394" height="116" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
-
-function DarkThemePersesLogo() {
-  return (
-    <svg width="367" height="120" viewBox="0 0 367 120" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <rect width="367" height="120" />
-      <path
-        d="M121.76 86V41.328H138.016C142.155 41.328 145.312 42.459 147.488 44.72C149.664 46.981 150.752 50.139 150.752 54.192C150.752 58.245 149.664 61.403 147.488 63.664C145.312 65.925 142.155 67.056 138.016 67.056H127.136V86H121.76ZM127.136 62.384H137.888C140.149 62.384 141.899 61.829 143.136 60.72C144.416 59.568 145.056 57.947 145.056 55.856V52.528C145.056 50.437 144.416 48.837 143.136 47.728C141.899 46.576 140.149 46 137.888 46H127.136V62.384ZM160.135 86V41.328H187.655V46H165.511V61.04H186.887V65.712H165.511V81.328H187.655V86H160.135ZM203.886 86H198.51V41.328H214.766C218.905 41.328 222.062 42.459 224.238 44.72C226.414 46.981 227.502 50.139 227.502 54.192C227.502 57.819 226.563 60.72 224.686 62.896C222.809 65.072 220.078 66.331 216.494 66.672L227.886 86H221.87L210.99 66.928H203.886V86ZM214.638 62.384C216.899 62.384 218.649 61.829 219.886 60.72C221.166 59.568 221.806 57.947 221.806 55.856V52.528C221.806 50.437 221.166 48.837 219.886 47.728C218.649 46.576 216.899 46 214.638 46H203.886V62.384H214.638ZM250.197 86.768C246.186 86.768 242.901 86.043 240.341 84.592C237.781 83.099 235.669 81.2 234.005 78.896L237.909 75.696C239.658 77.872 241.493 79.493 243.413 80.56C245.376 81.627 247.701 82.16 250.389 82.16C253.632 82.16 256.106 81.413 257.813 79.92C259.52 78.427 260.373 76.336 260.373 73.648C260.373 71.472 259.754 69.787 258.517 68.592C257.28 67.355 255.125 66.48 252.053 65.968L247.061 65.136C244.885 64.752 243.05 64.176 241.557 63.408C240.106 62.64 238.933 61.723 238.037 60.656C237.141 59.547 236.501 58.352 236.117 57.072C235.733 55.749 235.541 54.384 235.541 52.976C235.541 48.88 236.885 45.787 239.573 43.696C242.261 41.605 245.888 40.56 250.453 40.56C254.037 40.56 257.045 41.157 259.477 42.352C261.952 43.547 263.936 45.232 265.429 47.408L261.653 50.672C260.416 49.051 258.922 47.728 257.173 46.704C255.424 45.68 253.162 45.168 250.389 45.168C247.36 45.168 245.034 45.808 243.413 47.088C241.792 48.368 240.981 50.288 240.981 52.848C240.981 54.853 241.578 56.496 242.773 57.776C244.01 59.013 246.208 59.909 249.365 60.464L254.165 61.296C256.341 61.68 258.176 62.256 259.669 63.024C261.162 63.792 262.357 64.709 263.253 65.776C264.192 66.843 264.853 68.037 265.237 69.36C265.621 70.683 265.813 72.069 265.813 73.52C265.813 77.659 264.448 80.901 261.717 83.248C259.029 85.595 255.189 86.768 250.197 86.768ZM275.26 86V41.328H302.78V46H280.636V61.04H302.012V65.712H280.636V81.328H302.78V86H275.26ZM326.947 86.768C322.936 86.768 319.651 86.043 317.091 84.592C314.531 83.099 312.419 81.2 310.755 78.896L314.659 75.696C316.408 77.872 318.243 79.493 320.163 80.56C322.126 81.627 324.451 82.16 327.139 82.16C330.382 82.16 332.856 81.413 334.563 79.92C336.27 78.427 337.123 76.336 337.123 73.648C337.123 71.472 336.504 69.787 335.267 68.592C334.03 67.355 331.875 66.48 328.803 65.968L323.811 65.136C321.635 64.752 319.8 64.176 318.307 63.408C316.856 62.64 315.683 61.723 314.787 60.656C313.891 59.547 313.251 58.352 312.867 57.072C312.483 55.749 312.291 54.384 312.291 52.976C312.291 48.88 313.635 45.787 316.323 43.696C319.011 41.605 322.638 40.56 327.203 40.56C330.787 40.56 333.795 41.157 336.227 42.352C338.702 43.547 340.686 45.232 342.179 47.408L338.403 50.672C337.166 49.051 335.672 47.728 333.923 46.704C332.174 45.68 329.912 45.168 327.139 45.168C324.11 45.168 321.784 45.808 320.163 47.088C318.542 48.368 317.731 50.288 317.731 52.848C317.731 54.853 318.328 56.496 319.523 57.776C320.76 59.013 322.958 59.909 326.115 60.464L330.915 61.296C333.091 61.68 334.926 62.256 336.419 63.024C337.912 63.792 339.107 64.709 340.003 65.776C340.942 66.843 341.603 68.037 341.987 69.36C342.371 70.683 342.563 72.069 342.563 73.52C342.563 77.659 341.198 80.901 338.467 83.248C335.779 85.595 331.939 86.768 326.947 86.768Z"
-        fill="white"
-      />
-      <path
-        d="M66.625 24H24.375C20.8542 24 18 26.8542 18 30.375C18 33.8958 20.8542 36.75 24.375 36.75H66.625C70.1458 36.75 73 33.8958 73 30.375C73 26.8542 70.1458 24 66.625 24Z"
-        fill="#DE005D"
-      />
-      <path
-        d="M86.625 44.75H44.375C40.8542 44.75 38 47.6042 38 51.125C38 54.6458 40.8542 57.5 44.375 57.5H86.625C90.1458 57.5 93 54.6458 93 51.125C93 47.6042 90.1458 44.75 86.625 44.75Z"
-        fill="#DE005D"
-      />
-      <path
-        d="M66.625 65.5H24.375C20.8542 65.5 18 68.3542 18 71.875C18 75.3958 20.8542 78.25 24.375 78.25H66.625C70.1458 78.25 73 75.3958 73 71.875C73 68.3542 70.1458 65.5 66.625 65.5Z"
-        fill="#DE005D"
-      />
-      <path
-        d="M31.625 86.25H24.375C20.8542 86.25 18 89.1042 18 92.625C18 96.1458 20.8542 99 24.375 99H31.625C35.1458 99 38 96.1458 38 92.625C38 89.1042 35.1458 86.25 31.625 86.25Z"
-        fill="#DE005D"
-      />
-    </svg>
-  );
-}
-
-function PersesLogoCropped() {
-  return (
-    <svg width="120" height="120" viewBox="0 0 120 120" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M71.625 23H29.375C25.8542 23 23 25.8542 23 29.375C23 32.8958 25.8542 35.75 29.375 35.75H71.625C75.1458 35.75 78 32.8958 78 29.375C78 25.8542 75.1458 23 71.625 23Z"
-        fill="#DE005D"
-      />
-      <path
-        d="M91.625 43.75H49.375C45.8542 43.75 43 46.6042 43 50.125C43 53.6458 45.8542 56.5 49.375 56.5H91.625C95.1458 56.5 98 53.6458 98 50.125C98 46.6042 95.1458 43.75 91.625 43.75Z"
-        fill="#DE005D"
-      />
-      <path
-        d="M71.625 64.5H29.375C25.8542 64.5 23 67.3542 23 70.875C23 74.3958 25.8542 77.25 29.375 77.25H71.625C75.1458 77.25 78 74.3958 78 70.875C78 67.3542 75.1458 64.5 71.625 64.5Z"
-        fill="#DE005D"
-      />
-      <path
-        d="M36.625 85.25H29.375C25.8542 85.25 23 88.1042 23 91.625C23 95.1458 25.8542 98 29.375 98H36.625C40.1458 98 43 95.1458 43 91.625C43 88.1042 40.1458 85.25 36.625 85.25Z"
-        fill="#DE005D"
-      />
-    </svg>
-  );
-}
+import PersesLogoCropped from '../../components/logo/PersesLogoCropped';
+import DarkThemePersesLogo from '../../components/logo/DarkThemePersesLogo';
+import LightThemePersesLogo from '../../components/logo/LightThemePersesLogo';
 
 function AuthView() {
   const { isDarkModeEnabled } = useDarkMode();
   const isLaptopSize = useMediaQuery(useTheme().breakpoints.up('sm'));
   const authMutation = useAuthMutation();
+  const { successSnackbar, exceptionSnackbar } = useSnackbar();
   const [login, setLogin] = useState<string>('');
   const [password, setPassword] = useState<string>('');
+
+  const handleLogin = () => {
+    authMutation.mutate(
+      { login: login, password: password },
+      {
+        onSuccess: () => {
+          successSnackbar(`Successfully login`);
+        },
+        onError: (err) => {
+          exceptionSnackbar(err);
+        },
+      }
+    );
+  };
 
   return (
     <Container
@@ -129,7 +76,7 @@ function AuthView() {
         <Button
           variant="contained"
           disabled={authMutation.isLoading || login === '' || password === ''}
-          onClick={() => authMutation.mutate({ login: login, password: password })}
+          onClick={() => handleLogin()}
         >
           Login
         </Button>
